app/genai/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.2.2
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
loguru>=0.6.0
pydantic>=1.10.0
fastapi>=0.95.0
uvicorn>=0.20.0
streamlit>=1.25.0
plotly>=5.0.0
matplotlib==3.8.2
seaborn==0.13.0
tiktoken==0.5.2
faiss-cpu==1.7.4
scikit-learn>=1.1.0
requests>=2.28.0
sqlalchemy>=1.4.0
alembic>=1.8.0
jinja2>=3.1.0
aiofiles>=22.0.0
python-multipart>=0.0.5 